+++ src/frontend/src/lib/components/tokens/TokensDisplayHandler.svelte
@@ -2,8 +2,9 @@
 	import { debounce } from '@dfinity/utils';
 	import { combinedDerivedSortedNetworkTokensUi } from '$lib/derived/network-tokens.derived';
 	import { showZeroBalances } from '$lib/derived/settings.derived';
-	import type { TokenUi, TokenUiOrGroupUi } from '$lib/types/token';
+	import type { TokenUi } from '$lib/types/token';
 	import { groupTokensByTwin } from '$lib/utils/token-group.utils';
+	import type { TokenUiOrGroupUi } from '$lib/types/token-group';
 
 	// We start it as undefined to avoid showing an empty list before the first update.
 	export let tokens: TokenUiOrGroupUi[] | undefined = undefined;
