+++ src/frontend/src/lib/types/token-group.ts
@@ -1,0 +1,12 @@
+import type { TokenFinancialData, TokenId, TokenUi } from '$lib/types/token';
+
+//todo: separate typing from token id
+type GroupId = TokenId;
+
+export type TokenUiGroup = {
+	id: GroupId;
+	nativeToken: TokenUi;
+	tokens: TokenUi[];
+} & TokenFinancialData;
+
+export type TokenUiOrGroupUi = TokenUi | TokenUiGroup;
